+++ src/impl/socket/socket_com_server.c
@@ -103,14 +103,14 @@ static ComErcd Close( Com *pSuper )
     return COM_E_OK;
 }
 
-static ComErcd Read( Com *pSuper, char *pBuffer, int length )
+static ComErcd Read( Com *pSuper, char *pBuffer, size_t length )
 {
     SocketComServer *pSelf = (SocketComServer *)pSuper;
     CHECK_NULL( pSelf );
     CHECK_NULL( pBuffer );
     CHECK_SOCKFD( pSelf );
 
-    int totallen = 0;
+    size_t totallen = 0;
     while ( totallen < length ) {
 
         ssize_t recvlen = recv(
@@ -119,20 +119,20 @@ static ComErcd Read( Com *pSuper, char *pBuffer, int length )
             perror( "recv" );
             return COM_E_SYS;
         }
-        totallen += (int)recvlen;
+        totallen += (size_t)recvlen;
     }
 
     return COM_E_OK;
 }
 
-static ComErcd Write( Com *pSuper, const char *pBuffer, int length )
+static ComErcd Write( Com *pSuper, const char *pBuffer, size_t length )
 {
     SocketComServer *pSelf = (SocketComServer *)pSuper;
     CHECK_NULL( pSelf );
     CHECK_NULL( pBuffer );
     CHECK_SOCKFD( pSelf );
 
-    int totallen = 0;
+    size_t totallen = 0;
     while ( totallen < length ) {
 
         ssize_t sendlen = send(
@@ -141,7 +141,7 @@ static ComErcd Write( Com *pSuper, const char *pBuffer, int length )
             perror( "send" );
             return COM_E_SYS;
         }
-        totallen += (int)sendlen;
+        totallen += (size_t)sendlen;
     }
 
     return COM_E_OK;
